==> Notebook/05_Ttest.Rmd <==
---
title: "T de Student y sus variaciones"
output: html_notebook
---

# Introduccion

Prueba de hipotesis parametrica para comparar la media de una o dos poblaciones. Se usa cuando se tiene una variable dependiente y 1 variable independiente de 2 opciones.

```{r}
Fosf <- read.csv("~/CursoR/CursoRgit/Materiales/phosphorus.csv")

Profundidad <- read.csv("~/CursoR/CursoRgit/Materiales/IBPJ.Thaw.csv")

DOC <- read.csv("~/CursoR/CursoRgit/Materiales/Pond.Lake.DOC.csv")

Calidad_agua <- read.csv("~/CursoR/CursoRgit/Materiales/WQ.August.csv")
```

# Prueba de T de una muestra

Test de hipotesis estadistica para comparar la diferencia significativa entre una poblacion (tu variable) y un valor especifico

_¿Cuando usarla?_

* Analisis de gestion ambiental
* Control de calidad

```{r}
# checar normalidad

shapiro.test(Fosf$TDP)
hist(Fosf$TDP)

t.test(Fosf$TDP, mu = 75.6)
# no estan normalizados los datos por lo que se tienen que normalizar para realizar la prueba de "T":


t.test(log(Fosf$TDP), mu = log(75.6)) # mu valor de referencia

```

# Prueba de T pareada

La prueba de T pareada es un metodo que se usa para comprobar si la media entre pares de medidas es o no igual a cero.

_¿Cuando usarla?_

* Para pruebas de antes y despues (tienen que ser dos variables dependientes conectas entre si)

```{r}
# Checar normalidad en la diferencia entre cada set de datos

Dif <- Profundidad$Thaw.2010s - Profundidad$Thaw.1970s

shapiro.test(Dif)

par(mfrow=c(1,2))
hist(Dif)
qqnorm(Dif)

t.test(Profundidad$Thaw.1970s, Profundidad$Thaw.2010s,
       alternative=c("two.sided"),
       mu=0,paired=TRUE) 
```

## Ejercicio

Encontrar o crear un mini set de datos que podria ser analizado usando una prueba de t pareada. Realizar el test y concluir si existe o no una diferencia estadistica

_Nota_ codigo abs() es para valor absoluto, se usa cuando hay numeros negativos en nuestro set de datos

```{r}
datos <- data.frame(Agitacion_24_hrs = c(1,19,20,20,14,1), Agitacion_48_hrs = c(1,19,20,14,12,5), Agtacion_72_hrs = c(0,4,8,0,0,2))

Difer <- datos$Agitacion_24_hrs - datos$Agitacion_48_hrs

shapiro.test(Difer)

par(mfrow=c(1,2))
hist(Difer)
qqnorm(Difer)

t.test(datos$Agitacion_24_hrs, datos$Agitacion_48_hrs,
       alternative = c("two.sided"),
       mu = 0, paired = TRUE)

# no son significativamente diferentes, por lo que no hay diferencia entre los dos tiempos diferentes utilizados.

```

# Prueba de T para dos muestras

Esta es la mas comun o la mas conocida (tambien llamada prueba t de muestras independientes) es un metodo utilizado para probar si las medias de poblacion desconocidas de dos grupos son iguales o no.

De este tipo de prueba hay dos variables la T de student y el test de Welch's (no asume varianzas iguales)

_¿Cuando usarla?_

Comparar dos grupos diferentes en relacion con una variable dependiente:

* Biomasa de ua especie de planta en dos regimenes diferentes de crecimiento.
* Concentracion de metales pesados entres dos campos de cultivo.

```{r}

# Estructura:(variable dep ~ variable indep, data=tabla)

t.test(DOC ~ Habitat, data = DOC)

# Aqui si son significativaente dif
```

Como pueden observar la estructura del codigo de estas pruebas no es tan complicada como todo el codigo previo que ya hemos hecho. Sin embargo lo complicado a partir de aqui es usar criterio para ver CUAL de todas las pruebas se debe usar para analizar los datos que tengas.

## Ejercicio

Usando el data frame _calidad_Agua_ comparar DOC (ppm), fosforo reactivo soluble(SRP:ug/L), y clorofila en fitoplancton (ug/L) en estanques de Alaska entre los anios 1970s y 2010s.

```{r}
# 1. SRP
shapiro.test(Calidad_agua$SRP)

par(mfrow=c(1,2))
hist(Calidad_agua$SRP)
qqnorm(Calidad_agua$SRP)

t.test(log(SRP) ~ Era, data = Calidad_agua) # si hay diferencias significativas entres los anios y el fosforo reactivo soluble que hay en los estanques

# 2. Phyto
shapiro.test(Calidad_agua$Phyto)

par(mfrow=c(1,2))
hist(Calidad_agua$Phyto)
qqnorm(Calidad_agua$Phyto)

t.test(log(Phyto) ~ Era, data = Calidad_agua) # si hay diferencias significativas entres los anios y la clorofila en fitoplancton

# 3. DOC1
shapiro.test(Calidad_agua$DOC1)

par(mfrow=c(1,2))
hist(Calidad_agua$DOC1)
qqnorm(Calidad_agua$DOC1)

t.test(DOC1 ~ Era, data = Calidad_agua) # si hay diferencias significativas entre los anios y el carbono organico disuelto  

```
