==> Notebook/15_Herr_phyloseq.Rmd <==
---
title: "Herramientas para usar con Phyloseq"
output: html_notebook
---

```{r}
library(phyloseq)

# Paso 1: Subir tablas de dada
load("taxa_proyecto_Mariana.RData") # identficacion taxonomica
load("seq_conteos_proyecto_Mariana.RData") # abundancia ASVs

# Paso 2: metadatos
metadatos <- data.frame(Tratamiento = "Bioestimulante",
                        Suelo = "Salino")

row.names(metadatos) <- "S85"

# Paso 3: Renombrar muestras
row.names(seqtab.nochim) <- "S85"

# Paso 4: Cambiar nombres de secuencias
rando <- function(n = 5000) {
  a <- do.call(paste0, replicate(5, sample(LETTERS, n, TRUE), FALSE))
  paste0(a, sprintf("%04d", sample(9999, n, TRUE)), sample(LETTERS, n, TRUE))
}

dim(seqtab.nochim) # 5249
seqnames <- rando(5249) # entre parentesis va el resultado de la funcion anterior

# Guardar las secuencias en otro objeto antes de borrarlas
Secuencias <- colnames(seqtab.nochim)
write.csv(Secuencias, "~/CursoR/CursoRgit/Materiales/Secuencias_vid_s85.csv")

# Cambiando nombre con codigo
colnames(seqtab.nochim) <- seqnames

# Paso 5: Cambiar el nombre a las seuencias en taxa
row.names(taxa) <- seqnames

# Paso 6: PASO FINAL

vid_bio <- phyloseq(otu_table(seqtab.nochim,
                               taxa_are_rows = FALSE), 
                     sample_data(metadatos),
                     tax_table(taxa))

# y no se olviden de guardarlo
save(vid_bio, file = "Vid_S85.RData")

```

# Introduccion

Como vimos antes de irnos phyloseq nos ayuda a integrar todos nuestros datos en un objeto para poder analizarlo. La clase de hoy vamos a ver como phyloseq nos deja modificar estos objetos para visualizarlos mejor y seguir con los diferentes tipo de analisis usando los datos de vid.
 
```{r}
load("Vid_S85.RData") # siempre se empieza con el objeto de phyloseq

```

# Las primeras visualizaciones de nuestros datos

```{r}
# Redes de como estan interactuando las muestras
net <- make_network(vid_bio, "samples", max.dist=2)
plot_network(net, vid_bio, color = "Tratamiento", shape="Suelo",
             line_weight = 0.3, label=NULL)

plot_bar(vid_bio, fill = "Phylum")

plot_heatmap(vid_bio, taxa.label = "Phylum")
```

Si bien con estos datos ya podemos realizar grafico que nos acercan mas al analisis el hecho es de que siempre es comvenenientre pre-procesar los datos antes de cualquier grafico. Para ello phyloseq nos ofrece varias herramientas:

# Preproccesamiento de datos

## Filtraado

```{r}
# Porque tres muestras?

PS_filtered <- filter_taxa(vid_bio, # objeto 
                           function(OTU) sum(OTU) > 2, # condicion o funcion
                           TRUE) # cortar
# este es un proceso de filtrado por numero de muestras 
# de 5249 me dejo con 3700

# Remover taxa no identificada
PS_filtered <- subset_taxa(PS_filtered, # objeto 
                           !is.na(Phylum)) # condicion que se aplica para que de los que no tengan nada en phylum me los quite
# me dejo con 3605 
```

### Prune vs subset

```{r}
# usando datos de la tabla de taxa
Actino <- subset_taxa(vid_bio, # objeto
                      Phylum=="Actinobacteriota") # la condicion puede ser cualquier nivel de la jerarquia taxonomica

# usando abundancias quitar muestras
Actino <- prune_samples(sample_sums(Actino)>=50, # condicion que las que sean menor a 50 me las va a quitar
                        Actino) # objeto
```

# Union o merge

```{r}
# Uniendo muestras
Tratamientos <- merge_samples(vid_bio, # objeto
                              "Tratamiento") # condicion de metadatos

## Uniendo taxa
Actino_m <- merge_taxa(Acino, taxa_names(Actino)[1:5]) # Por numero

# Por jerarquia
PS_glom <- tax_glom(PS_filtered, # objeto
                     taxrank = "Genus", # nivel de la jerarquia
                     NArm=FALSE) # que no me quite los valores falsos
# uniendo phyloseqs
merge_phyloseq(Actino)
```

### Ejercicio unir las muestras de todos

```{r}
load("VidS81.RData")
V1 <- vid_bio

load("VS82.RData")
V2<- vid_bio

load("VidS83.RData")
V3 <- vid_bio

load("Vid_S85.RData")
V5 <- vid_bio

vid_bio <- merge_phyloseq(V1, V2, V3, V5)

save(vid_bio, file = "vid_ejemplos.RData")

load("vid_ejemplos.RData") # al usar merge fijarse que los datos coincidan

# Modificando la tabla de metadatos

vid_bio@sam_data$Suelo[is.na(vid_bio@sam_data$Suelo)] <- "Salino"
vid_bio@sam_data$Tratamiento[vid_bio@sam_data$Tratamiento == "Sin bioestimulante"] <- "Control"
vid_bio@sam_data <- vid_bio@sam_data[,1:2] # [filas,columnas]

save(vid_bio, file = "vid_ejemplos.RData")
load("vid_ejemplos.RData")
```

```{r}
load("phylo_soil.RData")
```

## Abundancia relativa

DADA2 nos da abundancias absolutas las cuales pasamos a nuestro objeto de phyloseq. La abundancia abdoluta es el conteo TOTAL de las especies e individuos dentro de cada especie

La abundancia relativa en cambio normaliza nuestras muestras para poder compararlas entre si. Para ello usa proporciones siendo 1 el 100% de nuestro conteo por muestras. De esta manera podemos comparar la composicion de nuestras muestras aunque no tengamos valores totales o absolutos iguales.

Para transformar de abundancia absoluta a relativa o cualquier otro tipo de calculo para nuestras muestras se usa la siguiente funcion en phyloseq:

```{r}
Psoil_rel <- transform_sample_counts(PS_filtered, # objeto a transformar
                                     function(x) x/sum(x)) # x es mi conteo
```

# Reexploramos graficos

```{r}
# 1. Graficos
Top_phyla <- tax_glom(Psoil_rel,taxrank = "Phylum", NArm = FALSE)
Top_phyla <- prune_taxa(names(sort(taxa_sums(Top_phyla), TRUE)[1:10]), #cond
                        Top_phyla) # objeto que voy a modificar

plot_bar(Top_phyla, fill = "Phylum")
plot_heatmap(Top_phyla, taxa.label = "Phylum")

```
