==> Notebook/16_Div_alfa.Rmd <==
---
title: "Diversidad alfa"
output: html_notebook
---

# Introduccion 

Como menciones anteriormente, la diversidad alfa es la diversidad media de especies en un sitio a escala local. Este es el primer nivel de comparacion de diversidad que van a usar para comparar sus muestras

```{r}
# Librerias
library(phyloseq)
library(ggplot2)
library(RColorBrewer)
library(tidyverse)
library(dplyr)
library(car)
library(breakaway) # chao-bunge
library(microbiome)
library(agricolae)

BiocManager::install("microbiome")

# Data
load("vid_ejemplos.RData")
```

# Indices de diversidad alfa

Calcular la diversidad alfaa en realidad es bastante sencillo. El reto para ustedes sera interpretar esos datos:

```{r}
# funcion de phyloseq
alphadiv <- estimate_richness(vid_bio, # objeto
                              split= TRUE, # default
                              measures = NULL) # elegir el tipo de indice, el default es NULL y te da como resultado 7 indices diferentes

# Separar metdatos para ponerlos en la tabla de div alfa
samplemeta <- as.data.frame(vid_bio@sam_data)
alphadiv <- cbind(samplemeta,alphadiv)
```

La funcion base de phyloseq calcula 7 metodos diferentes para diversidad alfa:

1. _Observed:_ cuenta el numero de especies por muestra (riqueza de especies)

2. _Chao1:_ Estimador de riqueza de especie bajo la hipótesis de que las especies raras dan la mayor información sobre aquellas que no se observaron. Chao1 es usado por lo regular para muestras pequeñas y es particularmente útil en datos sesgados a especies de baja abundancia.

3. _ACE:_ Abundance-based coverage estimator (ACE) es otro estimador no parametrico de riqueza de especie que usa la covertura de la muestra, basado en la suma de las probabilidades de las especies observadas

4. _Shannon H':_ es un indice que integra tanto riqueza como uniformidad. Sin entrar en las matematicas Shannon le pone mas enfasis en sus calculos a la riqueza de especies como componente y a la cobertura de especies raras.

5. _Simpson D1:_ Es una medida de diversidad que tiene en cuenta el número de especies presentes, así como la abundancia relativa de cada especie. A medida que aumentan la riqueza y la uniformidad de las especies, también aumenta la diversidad. El valor de D oscila entre 0 y 1. El índice de Simpson pone mayor énfasis en el componente de uniformidad y en los tipos de cobertura dominantes.

6. Dado que Simpson mide la probabilidad de que dos individuos seleccionados aleatoriamente de una muestra pertenezcan a la misma especie (o a alguna categoría distinta de especie) tambien se puede calcular _InvSimpson_

7. _Fisher:_ Este es un índice paramétrico de diversidad que supone que la abundancia de especies sigue la distribución de la serie logarítmica.

```{r}
estimate_richness(vid_bio, split = TRUE,
                  measures = c("Observed","Chao1","Shannon","Simpson")) # estas son las mas comunes a usarse
```

## Otros indices de diversidad que pueden calcular

```{r}
# del paquete breakaway
chao_bunge(vid_bio, cutoff = 10) # Similar a Chao1 pero mas sensible
# cuando usarlo? cuando tenga baja abundancia y gran porcentaje de muestras raras

# del paquete microbiome
Evenness <- evenness(vid_bio, # objeto
                     index = 'all', 
                     zeroes = TRUE, detection = 0) # default

# Unir y guardar nuestra tabla
alphadiv <- cbind(alphadiv, Evenness)
write.csv(alphadiv, "~/CursoR/CursoRgit/Materiales/Diversidad_alfa.csv")
```

De esta segunda seccion es importante notar que la funcion de evenness() calcula otros indices de diversidad alfa los cuales estan enfocados en la uniformidad de especies. Si notan la mayor parte de los indices anteriores le daban mas peso a la riqueza de especies por lo que el paquete microbiome se enfoca en otros tipos de indices. Como esta clase no es para explicarles cada uno de ellos de aqui el unico que les puede ser util en el futuro es:

8. _Pielou:_ Equidad (también conocida como equidad de Pielou) es la diversidad de Shannon dividida por el logaritmo del número de taxones. Esto mide la uniformidad con la que los individuos se dividen entre los taxones presentes.

[Mas info en indices de diversidad](https://www.nhm.uio.no/english/research/resources/past/help/diversity.html)

# Graficando diversidad alfa

```{r}
# Grafico usando la funcion de phyloseq

graf_alfa <- plot_richness(vid_bio, x= "Tratamiento", # var independiente 1
                            measures = c("Observed","Shannon","Simpson", "Chao1"),
                            color = "Suelo") + # var independiente 2
  scale_color_manual(values = c("#8B0A50", "#43CD80"),
                     name = "Suelo")+
  theme_bw()

graf_alfa$layers <- graf_alfa$layers[-1]
graf_alfa <- graf_alfa + geom_point(size=4, alpha=0.3)
graf_alfa

# A mano usando la tabla que creamos

## Pielou
ggplot(alphadiv, # objeto
       aes(x=Tratamiento,
           y=pielou, # y seria el indice
           color= Suelo))+
  geom_point(size=4, alpha=0.4)+
  scale_color_manual(values = c("#8B0A50", "#43CD80"))+
  theme_bw()
         
```

# Analisis estadistico

Como con los datos que vimos durante ese modulo, el analisis estadistio seguira el mismo flujo de trabajo segun la naturaleza de los datos

```{r}
## Primero revisamos normalidad ##

# shapiro test
for(i in 3:16){
  shapiro <- shapiro.test(alphadiv[,i])
  normal <- ifelse(shapiro[["p.value"]]>0.05, "YES", "NO")
  print(c(i,normal))
}

# histogramas si es que no eran normales
```

Dado que 

```{r}
# Revisar homogeneidad de varianza
for(i in 3:ncol(alphadiv)){ # empiezan en columna que terminan metadatos
  Lev_ex <- leveneTest(alphadiv[,i] ~ Suelo * Tratamiento,# var.indep
                       data = alphadiv)
  levene <- ifelse(Lev_ex[["Pr(>F)"]]>0.05, "YES", "NO")
  print(c(i,levene))
}

# Como solo tenemos una muestra no hay varianza por lo que no se puede calcular el test levene

# Observed

##ANOVA
ad_anova <- aov(Observed ~ Suelo * Tratamiento,
                data = alphadiv)
Anova(ad_anova)
# sale error porque estamos usando una sola muestra, si hubieran salido diferencias se procede con un tukey

ad_Tukey <- TukeyHSD(ad_anova)
```

Dado que solo tenemos " 1 muestra" por tratamiento no es correcto calcular un ANOVA ya que la ANOVA usa _grupos_. Lo ideal en este caso seria tratar de agrupar nuestras muestras por variables y llevar a cabo una T de student:

```{r}
t.test(Observed ~ Tratamiento, data = alphadiv) # no significativo
t.test(Chao1 ~ Tratamiento, data = alphadiv) # no significativo
t.test(Shannon ~ Tratamiento, data = alphadiv) # no significativo
t.test(pielou ~ Tratamiento, data = alphadiv) # no significativo
t.test(Simpson ~ Tratamiento, data = alphadiv) # salio significativo

# si hay diferencia en la diversidad de alfa segun el test de simpson, por lo que si hay diferencia en el suelo al usar el bioestimulante

```

Ya que esperamos que la aplicacion de nuestro bioestimulante mejor la diversidad alfa sin importar el tipo de suelo y basados en simpson si son diferentes

```{r}
# Pero comprobamos que no hay diferencias en tipo de suelo
t.test(Observed ~ Suelo, data = alphadiv) # NS
t.test(Chao1 ~ Suelo, data = alphadiv) # NS
t.test(Shannon ~ Suelo, data = alphadiv) # NS
t.test(pielou ~ Suelo, data = alphadiv) # NS
t.test(Simpson ~ Suelo, data = alphadiv) # NS

# Es bueno que no hayan diferencias en el suelo porque al descartalos como variable no afectamos el analisis
```

# Ejercicio final

```{r}
data("GlobalPatterns")
```

Van a usar los datos de "GlobalPatterns" que comparan las comunidades microbianas de 25 muestras ambientales y tres "comunidades simuladas" conocidas (un total de 9 tipos de muestras) a una profundidad promedio de 3.1 millones de lecturas por muestra. Con estos datos seguiran el flujo de trabajo que acabamos de ver y para el analisis estadistico compararan _"SampleType". Para este ejercicio solo analicen y grafiquen Chao1 y Shannon.

```{r}
# funcion de phyloseq
alphadiv <- estimate_richness(GlobalPatterns, 
                              split= TRUE, 
                              measures = c("Chao1","Shannon")) 

# Separar metdatos para ponerlos en la tabla de div alfa
samplemeta <- as.data.frame(GlobalPatterns@sam_data)
alphadiv <- cbind(samplemeta,alphadiv)

estimate_richness(GlobalPatterns, split = TRUE,
                  measures = c("Chao1","Shannon"))

# Grafico usando la funcion de phyloseq
graf_alfa <- plot_richness(GlobalPatterns, x= "SampleType", 
                            measures = c("Chao1","Shannon"),
                                         color = "SampleType")+
  scale_color_manual(values = c("#8B0A50", "maroon2","darkorange1", "#A020F0", "#00BFFF", "#CDCD00", "#008B45", "#00CD00", "#436EEE"),
                     name = "SampleType")+
                             theme_bw()
graf_alfa

## Primero revisamos normalidad ##

# shapiro test
for(i in 8:10){
  shapiro <- shapiro.test(alphadiv[,i])
  normal <- ifelse(shapiro[["p.value"]]>0.05, "YES", "NO")
  print(c(i,normal))
}

# Histograma
hist(alphadiv$Chao1)
hist(log(alphadiv$Chao1)) # no se puede normalizar por lo que se debe usar un test no parametrico Kruskal Wallis

hist(alphadiv$Shannon)
hist(sqrt(alphadiv$Shannon)) # si se normalizo por lo que se procede con anova

# Revisar homogeneidad de varianza
leveneTest(sqrt(Shannon) ~ SampleType, data = alphadiv) # pvalue= 0.048

##ANOVA
ad_anova <- aov(sqrt(Shannon) ~ SampleType,
                data = alphadiv)
Anova(ad_anova) # si hay diferencias significativas


ad_Tukey <- TukeyHSD(ad_anova)

# usando agricolae
summary(ad_anova)
rtout <- HSD.test(ad_anova, trt = "SampleType", alpha = 0.05)


```
