==> Notebook/17_Div_beta.Rmd <==
---
title: "Diversidad Beta"
output: html_notebook
---

```{r}
# Librerias
library(phyloseq)
library(ggplot2)
library(vegan) # analisis de diversidad beta
library(ggforce)
library(patchwork)

# Data
load("vid_ejemplos.RData")
load("Psoil_filt.RData")
```

# Introduccion

Mientras que la diversidad alfa representa la diversidad dentro de un ecosistema o una muestra, la diversidad beta representa la diferencia entre dos ecosistemas/muestras. En otras palabras, ¿qué tan similares o diferentes son dos ecosistemas o muestras? Entonces, la diversidad beta es una distancia entre dos muestras.

El análisis basado en distancia de la diversidad beta para microbiomas puede ser una herramienta poderosa para descubrir nuevas asociaciones entre la composición microbiana y una amplia variedad de fenotipos. 

Las medidas ecológicas de la diversidad beta tienen como objetivo capturar la diferencia global entre dos comunidades ecológicas. En el contexto del análisis de datos del microbioma, esto corresponde a diferencias entre sujetos en la composición microbiana. El análisis basado en la distancia o “a nivel de comunidad” luego compara estas diferencias por pares entre sujetos con diferencias por pares con respecto a algún fenotipo.

Estadisticamente, el analisis usando disimilitudes es mejor ya que evita la necesidad de realizar ajustes para una gran cantidad de comparaciones múltiples. Asimismo, los taxones no existen de forma aislada sino que tienen relaciones filogenéticas conocidas las cuales se pueden incluir en el analis. Otras características estructurales, como la similitud funcional, están bajo investigación y es posible que se puedan incorporar en el futuro.

# Tipos de distancias

Si creian que existen muchos metodos para analizar diversidad alfa, las cosas se complican aun mas con la diversidad beta. Phyloseq tiene integrados 44 tipos de distancias que se pueden calcular provenientes del paquete vegan (y ademas uno puede crear su propia formula):

```{r}
dist_methods <- unlist(distanceMethodList)
print(dist_methods)
```

__¿Pero entonces cual metodo elegimos para realizar nuestro analisis?__

Los ecologistas microbianos no utilizan distancias euclidianas, pero suelen utilizar distancias de Bray-Curtis, Jaccard o Unifrac de peso o sin peso para estimar la betadiversidad.

La disimilitud de _Bray-Curtis_ se basa en datos de ocurrencia (abundancia), mientras que la distancia de _Jaccard_ se basa en datos de presencia/ausencia (no incluye información de abundancia). Las distancias _UniFrac_ tienen en cuenta la tabla de ocurrencia y la diversidad de filogenia (distancia de secuencia). Distancias UniFrac ponderadas o no ponderadas dependiendo de si se tiene en cuenta la abundancia relativa o sólo la presencia/ausencia.
Las métricas de distancias están entre 0 y 1: 0 significa comunidades idénticas en ambas muestras y 1 significa comunidades diferentes en ambas muestras.

Top:

1. Bray-Curtis

2. Jaccard (no toma en cuenta abundancia)

3. Unifrac (solo si se tienen arboles filogeneticos)

¿Por qué las distancias de Bray-Curtis o Unifrac serían mejores para los datos de microbiota que la distancia euclidiana?

Los datos de microbiota son escasos y distancias específicas, como Bray-Curtis, Jaccard o las distancias Unifrac de peso/no peso, abordan mejor el problema de la presencia de muchos dobles ceros en los conjuntos de datos.

# Calcular distancias

```{r}
# Modificar phyloseq
Vid2 <- vid_bio

Vid2@sam_data <- Vid2@sam_data[,1]
#### Bray ####

bray_bdiv <- phyloseq::distance(Psoil_filt, # especificamos que es la funcion distance del paquete de phyloseq ya que hay otra funcion con ese nombre
                                method = "bray",
                                type = "sample") # sample es el default para que se calcule las distancias entre muestras y taxa es la otra opcion pero no la vamos a usar
```

## Mini ejercicio

Calculen las distancias con jaccard: Psoil y Vid2

```{r}
bray_bdiv <- phyloseq::distance(Psoil_filt,
                                method = "jaccard",
                                type = "sample")

bray_bdiv <- phyloseq::distance(Vid2,
                                method = "jaccard",
                                type = "sample")
```

# Ordinacion

Las medidas de diversidad beta se pueden visualizar utilizando varios métodos de ordinación. La ordinación es la técnicas para resumir y proyectar datos multidimensionales en un espacio de dimensiones inferiores (2-3d).

__Como el analisis de PCA__

## Metodos de ordinacion

Los metodos incluidos en phyloseq: c("DCA", "CCA", "RDA", "CAP", "DPCoA", "NMDS", "MDS", "PCoA")

1. DCA (el default): Performs detrended correspondence analysis usingdecorana

2. CCA Analisis de correspondencia (a.k.a. canonical correspondence analysis).

3. RDA Analisis de Redundancia (equivalente al PCA)

4. CAP Constrained Analysis of Principal Coordinates or distance-based RDA

5. DPCoA Doble Analisis de coordenadas principales usa distancia filogenetica como correcion

6. __NMDS__ Escalamiento multidimensional no métrico. NMDS maximiza la correlación de orden de clasificación entre muestras y tambien puede usarse con datos no parametricos. La mejor manera de visualizar la diversidad beta, o cuán diferentes son las muestras entre sí, es mediante escalamiento multidimensional no métrico (nMDS). Esto es similar al análisis de coordenadas principales o PCA/PCoA si ha oído hablar de eso, solo que nMDS es más robusto estadísticamente con múltiples iteraciones en la forma de la parte trymax del comando.

7. __MDS/PCoA__ Realiza un análisis de coordenadas principales (también llamado descomposición de coordenadas principales, escalamiento multidimensional (MDS) o escalamiento clásico) de una matriz de distancias 

```{r}
bray_ord <- ordinate(Psoil_filt, # objeto
                     method = "NMDS", # metodo
                     distance = bray_bdiv)
```

Entonces cual elegir?

Se debe elegir PCoA si las distancias entres muestras son tan cercanas que una transformacion lineal seria suficiente. Mientras tanto, se recomiend NMDS para resaltar la estructura de gradiente dentro de los datos. Otro metodo de ordenacion comun es el PCA, que es simplemente un tipo de PCoA que utiliza la distancia euclidiana. 

_NMDS_ > _PCoA_

La mejor manera de visualizar la diversidad beta, o cuan diferentes son las muestras entre si, es mediante escalamiento multidimensional no metrico (nMDS). Esto es similar al analisis de coordenadas principales o PCA/PCoA si ha oido hablar de eso, solo que NMDS es mas robusto estadisticamente con multiples iteraciones.

## Mini ejercicio

Calculen la ordinacion con PCoA y tambien calculenla para las distancias jaccard que teniamos.

```{r}
jac_bdiv <- phyloseq::distance(Psoil_filt,
                                method = "jaccard",
                                type = "sample")

bray_bdiv <- phyloseq::distance(Psoil_filt,
                                method = "bray",
                                type = "sample")

bray_ord_PCoA <- ordinate(Psoil_filt,
                     method = "PCoA", 
                     distance = bray_bdiv)

jaccard_ord <- ordinate(Psoil_filt,
                     method = "NMDS", 
                     distance = jac_bdiv)

jaccard_ord_PCoA <- ordinate(Psoil_filt,
                     method = "PCoA", 
                     distance = jac_bdiv)

# estos valores obtenidos de la ordinacion no se interpretan pero se utilizan para graficar
```

# Visualizacion

```{r}
p_bray <- plot_ordination(Psoil_filt, # objeto
                          bray_ord, # distancia bray con nmds
                          "samples",
                          color = "Plant_Type",
                          shape = "Treatment") +
  scale_color_manual(values = c("burlywood3","steelblue","lightpink"),
                     name = "Plant Type") +
  theme_bw()+
  geom_point(size=3.8)

p_bray

# Modificando las figuras
p_bray <- p_bray +
  scale_shape_manual(values = c(15,16,17,18),
                     labels = c("Pi", "Low P", "Phi", "Pi/Phi mix"))
p_bray

# Agregando elipse

p_bray <- p_bray +
  geom_mark_ellipse(aes(fill= Plant_Type,
                        color= Plant_Type), # tiene que ser igual que arriba
                    expand = c(0.0035)) + # ancho del circulo
  scale_fill_manual(values = c("burlywood3","steelblue","lightpink"),
                    name = "Plant Type")
p_bray

# o agregando poligono
p_bray <- p_bray +
  geom_polygon(aes(fill= Plant_Type),
                    alpha = c(0.2)) + # ancho del circulo
  scale_fill_manual(values = c("burlywood3","steelblue","lightpink"),
                    name = "Plant Type")

```

# Ejercicio

Van hacer otros tres graficos:

* El primero con distancias de bray pero ordinacion PCoA

* El segundo con Jaccard y ordinacion NMDS

* El tercero Jaccard y ordinacion PCoA

```{r}
# El primero con distancias de bray pero ordinacion PCoA

p_bray_PCoA <- plot_ordination(Psoil_filt, # objeto
                          bray_ord_PCoA, # distancia bray con PCoA
                          "samples",
                          color = "Plant_Type",
                          shape = "Treatment") +
  scale_color_manual(values = c("burlywood3","steelblue","lightpink"),
                     name = "Plant Type") +
  theme_bw()+
  geom_point(size=3.8)

p_bray_PCoA <- p_bray_PCoA +
  scale_shape_manual(values = c(15,16,17,18),
                     labels = c("Pi", "Low P", "Phi", "Pi/Phi mix"))

p_bray_PCoA <- p_bray_PCoA +
  geom_mark_ellipse(aes(fill= Plant_Type,
                        color= Plant_Type), # tiene que ser igual que arriba
                    expand = c(0.0035)) + # ancho del circulo
  scale_fill_manual(values = c("burlywood3","steelblue","lightpink"),
                    name = "Plant Type")

p_bray_PCoA

# El segundo con Jaccard y ordinacion NMDS

p_jac <- plot_ordination(Psoil_filt, # objeto
                          jaccard_ord, # distancia jac con NMDS
                          "samples",
                          color = "Plant_Type",
                          shape = "Treatment") +
  scale_color_manual(values = c("burlywood3","steelblue","lightpink"),
                     name = "Plant Type") +
  theme_bw()+
  geom_point(size=3.8)

p_jac <- p_jac +
  scale_shape_manual(values = c(15,16,17,18),
                     labels = c("Pi", "Low P", "Phi", "Pi/Phi mix"))

p_jac <- p_jac +
  geom_mark_ellipse(aes(fill= Plant_Type,
                        color= Plant_Type), # tiene que ser igual que arriba
                    expand = c(0.0035)) + # ancho del circulo
  scale_fill_manual(values = c("burlywood3","steelblue","lightpink"),
                    name = "Plant Type")

p_jac

# El tercero Jaccard y ordinacion PCoA

p_jac_PCoA <- plot_ordination(Psoil_filt, # objeto
                          jaccard_ord_PCoA, # distancia jac con NMDS
                          "samples",
                          color = "Plant_Type",
                          shape = "Treatment") +
  scale_color_manual(values = c("burlywood3","steelblue","lightpink"),
                     name = "Plant Type") +
  theme_bw()+
  geom_point(size=3.8)

p_jac_PCoA <- p_jac_PCoA +
  scale_shape_manual(values = c(15,16,17,18),
                     labels = c("Pi", "Low P", "Phi", "Pi/Phi mix"))

p_jac_PCoA <- p_jac_PCoA +
  geom_mark_ellipse(aes(fill= Plant_Type,
                        color= Plant_Type), # tiene que ser igual que arriba
                    expand = c(0.0035)) + # ancho del circulo
  scale_fill_manual(values = c("burlywood3","steelblue","lightpink"),
                    name = "Plant Type")

p_jac_PCoA

(p_bray|p_bray_PCoA)/(p_jac|p_jac_PCoA)
```

En realidad no habian diferencias en los graficos podemos ver que los patrones se mantienen

# Analisis estadisco multivariable

## PERMANOVA

Para probar si los grupos son diferentes con respecto al centroide y la dispersión, se realizará una prueba estadística PERMANOVA. Para ello se utilizará una extensión multivariada de ANOVA, ya que hay muchas OTU que se utilizarán en la prueba. La extensión se basa en distancias entre muestras. La prueba compara distancias de muestras dentro del mismo grupo con distancias de muestras de diferentes grupos. Si la distancia entre muestras de diferentes grupos es mucho mayor que las muestras del mismo grupo, concluimos que los grupos no son iguales.

Para probar la importancia del resultado, se utiliza una prueba de permutación. Por lo tanto, todas las muestras se mezclan aleatoriamente en los grupos y la prueba se repite muchas veces. Si la relación (entre la distancia del grupo/la distancia dentro del grupo) es mucho mayor para los datos originales que para las permutaciones, concluimos que hay una diferencia estadísticamente significativa.

La prueba se puede aplicar en combinación con cualquier medida de distancia.

```{r}
### Extraer data frames del objeto phyloseq ###

# Count table: ASVs
SoilASVs <- as.data.frame(Psoil_filt@otu_table)

# metadata or environment
SoilMeta <- as.matrix(Psoil_filt@sam_data)
SoilMeta <- as.data.frame(SoilMeta)

#Adonis es el codigo de la permanova
adonis2(SoilASVs ~ Treatment * Plant_Type,
        data = SoilMeta, # de donde vienen los metadatos?
        permutations = 999, # este numero es arbitrario
        method = "bray") # mismo a las distancias que calculamos

adonis2(SoilASVs ~ Treatment * Plant_Type,
        data = SoilMeta, 
        permutations = 10000, # este numero es arbitrario
        method = "jaccard", binary=TRUE) # binary significa ausencias/presencia
```
La PERMANOVA es muy sensible anumeros bajos de muestreo

### Variables opcionales

Estratificacion: Entre que grupos se haran las comparaciones

```{r}
## By plant type
adonis2(SoilASVs ~ Treatment,
        strata = SoilMeta$Plant_Type, # separar entre tipo de planta
        data = SoilMeta, 
        permutations = 999, # este numero es arbitrario
        method = "bray")

## By treatment
adonis2(SoilASVs ~ Plant_Type, # la variable  analizar
        strata = SoilMeta$Treatment, # los grupos por los que vas analizar
        data = SoilMeta, 
        permutations = 999, # este numero es arbitrario
        method = "bray")

```

Si el P_value es inferior a 0.05 por lo que podemos concluir que la intravariabilidad es menor que la intervariabilidad entre los dos itios y que tenemos un efecto significativo

## ANOSIM

Es la version no parametrica de un PEMANOVA. ANOSIM evalua si las distancias o diferencias entro del grupo son mayores o iguales a las distancias entre grupos utilizando los rangos de todas las distancias de muestra por pares. Es menos robusto que la PERMANOVA pero es un analisis mas simplificado.

```{r}
anosim(SoilASVs, # Tabla de conteo
       SoilMeta$Plant_Type, # metadatos
       distance = "bray",
       permutations = 999)

anosim(SoilASVs, # Tabla de conteo
       SoilMeta$Treatment, # metadatos
       distance = "bray",
       permutations = 999) # no hay significancias

# (r) entre mas cerca de 1 es mejor, mas cerca a 0 estan mayormente distribuidos
```

Al interpretar estos resultados, desea observar la estadistica R de ANOSIM y los valores de significancia. Un valor de significancia inferior a 0.05 generalmente se considera estadisticamente significativo y significa que s epuede rechazar la hipotesis nula. Por lo tanto, existe una diferencia estadisticamente significativa en las comunidades microbianas entre sus grupos. Mayor que 0.05 significa que no hay diferencia estadistica entre las comunidades microbianas de su grupos.

" +La estadistica "R" de ANOSIM compara la media de las disimilitude clasificadas dentro de los grupos. Un valor de R cercano a "1" sugieres disimilitud entre grupos, mientras que un valor de R cercano a "0" sugiere una distribucion uniforme de rangos altos y bajos dentro y entre grupos". En otras palabras, cuanto mayor sea el valor R, más diferentes serán sus grupos en términos de composición de la comunidad microbiana.

## Mini ejercicio

Repetir este analisis estadistico con jaccard

```{r}
anosim(SoilASVs, # Tabla de conteo
       SoilMeta$Plant_Type, # metadatos
       distance = "jaccard",
       permutations = 999)

anosim(SoilASVs, # Tabla de conteo
       SoilMeta$Treatment, # metadatos
       distance = "jaccard",
       permutations = 999)
```

# Ejercicio final

Realizar analisis de div beta comparando la variable "SeqTech" como grupos usando "bray"

```{r}
data("enterotype")

# distancia y ordinacion
bray_bdiv_ejer <- phyloseq::distance(enterotype,
                                method = "bray",
                                type = "sample")

bray_ord_ejer <- ordinate(enterotype,
                     method = "NMDS", 
                     distance = bray_bdiv_ejer)

# grafico
p_bray_ejer <- plot_ordination(enterotype,
                          bray_ord_ejer, # distancia bray con nmds
                          "samples",
                          color = "SeqTech")+
  scale_color_manual(values = c("#43CD80","mediumpurple","#FF8C69"),
                     name = "SeqTech") +
  theme_bw()

p_bray_ejer <- p_bray_ejer +
  geom_mark_ellipse(aes(fill= SeqTech,
                        color= SeqTech), # tiene que ser igual que arriba
                    expand = c(0.0035)) + # ancho del circulo
  scale_fill_manual(values = c("#43CD80","mediumpurple","#FF8C69"),
                    name = "SeqTech")
p_bray_ejer

# PERMANOVA
enterotypeASVs <- as.data.frame(t(enterotype@otu_table)) # pusimos transpose para modificar de filas a columnas para que las observaciones fueran las mismas

EnterotypeMeta <- as.matrix(enterotype@sam_data)
EnterotypeMeta <- as.data.frame(EnterotypeMeta)


adonis2(enterotypeASVs ~ SeqTech,
        data = EnterotypeMeta, 
        permutations = 10000, 
        method = "bray")

#ANOSIM
anosim(enterotypeASVs, 
       EnterotypeMeta$SeqTech, 
       distance = "bray",
       permutations = 999)

# En este caso R si es alto por lo que mis grupos si son diferentes
```
