==> Notebook/18_Analisis_tax.Rmd <==
---
title: "Análisis de composición taxonómica"
output: html_notebook
---

# Introducción 


Hasta este momento hemos estado contando CUANTO hay dentro de nuestras muestras pero lo verdaderamente útil es saber QUË hay dentro de las muestras. Esto es a lo que se le llama _analisis taxonómico._

__¿Por qué es importante?__

Como les comentaba las comunidades microbianas en suelos son muy dinamicas y muchas veces (como con mis datos) no ves cambios en las cantidades totales de microorganismos. Sin embargo, al ver la taxonomia las abundancias relativas de algunos grupos pueden ser modificados ya sea aumentado o disminuido. 

De esta manera se espera que cualquier producto o tratamiento que apliquemos al suelo, no tenga un efecto detrimental en grupos taxonomicos que son buenos para la planta. Dado que todas estos hongos y microbios interactuan constantemente en este microecosistema estos cambios en abundancia pueden darnos pista de como interactuan.

```{r}

## Librerías ##

library(phyloseq)
library(tidyverse)
library(dplyr)
library(RColorBrewer)
library(pheatmap) # Heatmaps
library(microbiome)
library(ggsignif)
library(scales) # Modificar escalas
library(car)
library(wesanderson)


# DATA 

load("ITS_filt.Rdata")
load("vid_ejemplos.RData")

```

Este analisis no cumple con un proceso en específico y en realidad es exploratorio. Para ello necesitamos visualizar nuestros datos, o sea, hacer variedad de gráficos y de ahí se elige a cuales grupos vale la pena hacer análisis taxonómico 

# 1. Visualizar a nivel de Phylum

## Heatmap

```{r}

Phyla_fun <- tax_glom(vid_bio, taxrank = "Phylum", NArm = FALSE)

## Extraer datos del objeto phyloseq

OTU_matrix <- as.data.frame(Phyla_fun@otu_table)
Tax_matrix <- as.data.frame(Phyla_fun@tax_table)

# Renombramos las columnas de nuestras abundancias con el phylum de la tabla taxa

colnames(OTU_matrix) <- Tax_matrix$Phylum 

# Quitar phylum desconocido (SOLO DATOS DE VID). Se quitaron las columnas con datos NA

OTU_matrix <- OTU_matrix[, -c(10,21)]

# crear matriz a partir de OTU siempre se debe usar en vez de tabla porque solo necesitamos valores numericos
Phyla_matrix <- as.matrix(t(OTU_matrix)) # las muestras deben quedar como columnas y las filas como phylum por eso se utiliza transpose(t)

```

# Procesamiento de datos 

Este proceso es específico para cada set de datos. Usualmente los heatmaps siempre van de mayor a menor, así que ese paso siempre se realiza, pero el resto consta de de ordenar las muestras y renombrar variables.

```{r}

Phyla_matrix <- Phyla_matrix[order(Phyla_matrix[,1], # se escoge la columna 1para que me empiece a ordenar de mayor a menor
                                   decreasing = TRUE),] # Ordenar de mayor a menor 

# Cambiar el orden de las muestras

sorder <- c("S81", "S85", "S82", "S83")

Phyla_matrix <- Phyla_matrix[,sorder]

row.names(Phyla_matrix)[16] <- "SAR324 Clade" # Con este le cambiamos el nombre a una fila que tenía un nombre muy largo 

```

## Usando el paquete Pheatmap para realizar heatmaps de análisis taxonómico 

```{r}

pheatmap(Phyla_matrix) # El objeto siempre debe de estar en formato de matriz y los valores como numéricos

```

### _Notan algo en el orden de nuestras filas y columnas? Clusters_

La funcion default de pheatmap es que automaticamente te reacomoda (cluster) tus filas y columnas de acuerdo a como cree que estan mas relacionadas las variables; es decir las agrupa. Esta funcion suele ser util cuando haces heatmaps de genes ya que te permite ver como se activan o apagan en conjunto y si hay "clusters" de genes relacionados. Pero en nuestro caso arruino todo el pre-procesamiento previo que nosotros le dimos por lo que tenemos que quitarle ese default.

```{r}
pheatmap(Phyla_matrix,
         cluster_rows = FALSE, cluster_cols = FALSE) # quitar clusters

```

### _El siguiente problema? La escala_

Si se fijan los colores en este heatmap son cero utiles. Solo tenemos cambios graduales en una columna y el resto se ven todas horribles. Esto ocurre porque nuestros datos se puede arreglar de varias maneras y la funcion de pheatmap tiene el argumento de scale para tratar de solucionarlo. El default siempre es _none_ pero vamos tambien puede modificarlo por filas y columnas:

```{r}
pheatmap(Phyla_matrix,
         cluster_rows = FALSE, cluster_cols = FALSE,
         scale = "none") # es el default

pheatmap(Phyla_matrix,
         cluster_rows = FALSE, cluster_cols = FALSE,
         scale = "column") # transforma los valores usando de escala las muestras. Casi nunca lo vamos a usar al menos que tengamos muchas muestras de una sola

pheatmap(Phyla_matrix,
         cluster_rows = FALSE, cluster_cols = FALSE,
         scale = "row") # transforma los valores usando de escala los phylum y muestra las diferencias entre las muestras y un phylum no entre los phylums
```

Cual creen que es mas util?

Desafortunadamente ninguno de ellos es muy util. El de columnas no hizo nada porque el problema recae en que tenemos phylums con muuucha abundancia y otros que casi estan llenos de ceros mientras; pero eso pasa en todas las muestras por lo que al tratar de escalarlo se ve asi. 

En el caso de la transformacion por filas nos va un poco mejor porque trata de normalizar las abundancias de todos los phylums a una escala. Sin embargo, por el mismo problema anterior lo unico que nos resalta es si alguna de las muestras tiene numeros extranamente altos o bajos DENTRO de ese mismo phylum. Es decir nos compara dentro de phylums pero no entre ellos. Asi que les voy a compartir el codigo que me tomo literalmente 1 semana encontrar y basicamente salvo mi tesis.

```{r}
## Funcion especificamente para separar nuestros datos en secciones

quantile_breaks <- function(xs, n = 10) {
  breaks <- quantile(xs, probs = seq(0, 1, length.out = n))
  breaks[!duplicated(breaks)]
} # recuerden que la funcion nunca se cambia de nombre(copien y peguen siempre entre documentos)

mat_breaks <- quantile_breaks(Phyla_matrix, # es la matriz
                              n = 10) # el default es 10 pero se puede poner el numero de cortes que quiero a mis datos

# Volvemos al heatmap

pheatmap(Phyla_matrix,
         cluster_rows = FALSE, cluster_cols = FALSE,
         scale = "none", # regresamos a que no nos ponga escala
         breaks = mat_breaks) # aqui ponemos el objeto con cortes que creamos
```

### _¿Porque no se ven bien los colores?_

Aunque ya pueden ver mejor en nuestra escala el problema ahora es que la funcion no tiene la cantidad adecuada de colores para representar nuestros datos. Por ello nosotros le vamos a dar otros colores:

```{r}
pheatmap(Phyla_matrix,
         cluster_rows = FALSE, cluster_cols = FALSE,
         scale = "none", # regresamos a que no nos ponga escala
         breaks = mat_breaks,
         color = colorRampPalette(c("#EEEEE0", "slategray3", "#8B668B", "#551A8B"))(9)) # este numero debe ser mas pequeño que n de la funcion
```

#### Mini ejercicio

Cda quien va a poner un numero diferentes en los breaks como cambia el grafico

```{r}
mat_breaks <- quantile_breaks(Phyla_matrix, 
                              n = 30)

pheatmap(Phyla_matrix,
         cluster_rows = FALSE, cluster_cols = FALSE,
         scale = "none", 
         breaks = mat_breaks,
         color = colorRampPalette(c("snow2", "#B6AAE1", "mediumpurple3", "#0E021A"))(25))

pheatmap(Phyla_matrix,
         cluster_rows = FALSE, cluster_cols = FALSE,
         scale = "none", 
         breaks = mat_breaks,
         color = colorRampPalette(c("snow2","lightskyblue1","plum3","magenta4","steelblue4","black"))(25))

# No hubo diferencia al aumentar mas breaks
```

```{r}
# a partir de aqui va a ser diferente para cada heatmap
pheatmap(Phyla_matrix,
         cluster_rows = FALSE, cluster_cols = FALSE,
         scale = "none", 
        color = colorRampPalette(c("snow2","lightskyblue1","plum3","magenta4","steelblue4","black"))(25),
         breaks = mat_breaks,
         fontsize = 10, # tamaño de letra
         gaps_col = c(2), # separaciones entre columnas de bioestimulantes y control
         labels_col = c("Tratamiento salino", "Tratamiento no salino","Control salino", "Control no salino")) # nombres de columnas
```

Otros posibles cambios que le podemos hacer a nuestro heatmap:

* border_color: colores de las celdas

* cellwidth/celheight: ancho/alto de celda en puntos; si no se pone los valores depende del tamano de la ventana.

* clustering_distance_rows o columns: la distancia para clusterizar

* clustering_method: metodos en hclust

* cutree_rows/cols: numero de clusters en el que los datos se van a dividir

* treeheight_row/col: altura de los arboles de clusters default es 50 puntos

* legend: true or false

* legend_breaks: vector con los cortes de la leyenda (colores de leyenda)

* legend_labels: vector con los nombres de los cortes para a leyenda

* annotation_row/col: filas de colores extra para demostrar los metadatos en el heatmap.

* annotation_colors: colores manuales de las anotaciones

* annotation_legend: true or false

* annotation_names_row/col: true or false se veran los nombres de las anotaciones o no

* show_rownames/colnames: true or false

* main: titulo del grafico

* fontsize_row/col: cambiar el tamano de letra por separado

* angle_col: angulo del texto. Solo se pueden elegir de estas opciones: (0, 45, 90, 270 and 315)

* display_numbers: true or false. Poner los espacios en el heatmap.

* number_format/color

* fontsize_number

* gaps_row/col: vector de donde poner los espacios en el heatmap

* filename: camino para guardar el heatmap como imagen

* width/height: en pulgadas

* na_col: especificar el color de las celdas sin color

### Mini ejercicio:

Agreguenle o modifiquen 2 o mas de estos argumentos en su grafico

```{r}
pheatmap(Phyla_matrix,
         cluster_rows = FALSE, cluster_cols = FALSE,
         scale = "none", 
         color = colorRampPalette(c("snow2","lightskyblue1","plum3","magenta4","steelblue4","black"))(25),
         breaks = mat_breaks,
         fontsize = 10, 
         gaps_col = c(2), 
         labels_col = c("S85", "S81","S83", "S82"),
         angle_col = 0,
         fontsize_col = 10,
         fontsize_row = 8,
         main = "Análisis taxonómico por Phylum")
```

### Ejercicio final

Usando global patterns creen un heatmap de los phylums enontrados

```{r}
data("GlobalPatterns")

Phyla_fun <- tax_glom(GlobalPatterns, taxrank = "Phylum", NArm = FALSE)

OTU_matrix <- as.data.frame(Phyla_fun@otu_table)
Tax_matrix <- as.data.frame(Phyla_fun@tax_table)

Tax_matrix[54,2] <- "Unknown"
row.names(OTU_matrix) <- Tax_matrix$Phylum 

Phyla_matrix <- as.matrix(OTU_matrix)

class(Phyla_matrix) <- "numeric" # para usar la funcion los datos siempre deben ser numericos o sea estar en una matriz

Phyla_matrix <- Phyla_matrix[order(Phyla_matrix[,1], 
                                   decreasing = TRUE),]

quantile_breaks <- function(xs, n = 10) {
  breaks <- quantile(xs, probs = seq(0, 1, length.out = n))
  breaks[!duplicated(breaks)]
}

Phyla_matrix <- Phyla_matrix[-38,] # para eliminar una columna que tiene NA

mat_breaks <- quantile_breaks(Phyla_matrix,
                              n = 10)

pheatmap(Phyla_matrix,
         cluster_rows = FALSE, cluster_cols = FALSE,
         scale = "none", 
         color = colorRampPalette(c("snow2", "#B6AAE1", "mediumpurple3", "#0E021A"))(7),
         breaks = mat_breaks,
         fontsize = 10,
         angle_col = 315,
         fontsize_col = 8,
         fontsize_row = 6,
         main = "Análisis taxonómico por phylum")
```

## Grafico de Barras

Se acuerdan que al ver las funciones que nos daba phyloseq vimos como hacer graficos de barras? Pues bueno ahora vamos a aprender a hacerlos bonitos:

### Pre-procesamiento de datos

```{r}
# Recuerden que este tipo de graficos siempre se hacen con abundancias relativas.

Psoil_rel <- transform_sample_counts(vid_bio, # objeto a transformar
                                   function(x) x/sum(x)) # x es mi conteo

# Otra vez juntar phylums
Top_phyla <- tax_glom(Psoil_rel, taxrank = "Phylum", NArm = FALSE)

# Seleccionar solo los valores top
Top_phyla <- prune_taxa(names(sort(taxa_sums(Top_phyla), TRUE)[1:10]),
                        Top_phyla)
```

### El grafico

Usando la funcion plot_composition del paquete microbiome

```{r}
# Grafico base
plot_composition(Top_phyla, # objeto en valores relativos
                 plot.type = "barplot", # tipo de grafico tambien hace heatmap
                 )

# Ahora mas bonito

plot_composition(Top_phyla,plot.type = "barplot")+
  theme_bw()+
  scale_fill_brewer(palette = "RdYlGn", name = "Phylum",
                    labels=c("Actinobacteriota", "Chloroflexi","Proteobacteria", "Acidobacteriota", "Verrumcomicrobiota", "Bacteroidota", "Cyanobacteria","Patescibacteria", "Planctomycetota","Firmicutes"))+
  guides(x=guide_axis(angle=90))+
  scale_x_discrete(labels=c("Tratamiento salino", "Control no salino", "Control salino","Tratamiento salino"))
```

Si bien este grafico ya es hermoso, la forma mas simple de representar abundancia relativa es con porcentajes. Asi que les voy a ensenar la formula de scales te modifica etso facilmente

```{r}
plot_composition(Top_phyla,plot.type = "barplot")+
  theme_bw()+
  ggtitle("Abundancias relativas a nivel Phylum")+
  theme(plot.title = element_text(hjust = 0.5, size = 15, face = "bold"))+
  xlab("Muestras")+
  ylab("Abundancia")+
  scale_fill_brewer(palette = "Spectral", name = "Phylum",
                    labels=c("Actinobacteriota", "Chloroflexi","Proteobacteria", "Acidobacteriota", "Verrumcomicrobiota", "Bacteroidota", "Cyanobacteria","Patescibacteria", "Planctomycetota","Firmicutes"))+
  guides(x=guide_axis(angle=0))+
  scale_x_discrete(labels=c("S81", "S82", "S83","S85"))+
  scale_y_continuous(breaks = pretty_breaks(n = 10),# numero de separaciones
                     labels = scales::percent)# representadas en %
```

Usando estos dos graficos que pueden conclur de sus datos?
Hay algun phylum que crean que vale la pena revisar a detalle? 

Proteocateria y actinobacteriota por abundancias
Cyanobacteria por diferencias entre tratamientos

Si bien ya les ensene a presentar graficos bonitos y definitivamente cualquiera de estos graficos podrian incluirlo en un reporte, la visualizacion tambien nos permite darnos cuenta si existen ciertos patrones en abundancia. De esta manera podemos seleccionar alguno de estos grupos y verlos en la lupa de una manera mas minuciosa.

Por ejemplo, yo de mis datos podria decir que dada la dominancia que tiene Ascomycota tal vez vale la pena revisarla minuciosamente. Sin embargo, el heatmap no nos mostro si habian diferencias entre plantas o tratamientos (porque tal vez no las hay?) por lo que hacer varios loops de analisis estadistico seria muy util para elegirlos

# Analizar estadisticamente a nivel de phylum

El ejemplo va a ser con ANOVA y por eso estamos usando los datos de ITS
Para los graficos ya habiamos extraido nuestras tablas de conteos y taxa, pero ahora tambien necesitamos nuestros metadatos:

## Preprocesamiento:
```{r}
Phyla_fun <- tax_glom(ITS_filtered, taxrank = "Phylum", NArm = FALSE)

## Extraer datos del objeto phyloseq

OTU_matrix <- as.data.frame(Phyla_fun@otu_table)
Tax_matrix <- as.data.frame(Phyla_fun@tax_table)

## Extract metadata
metadata <- as.data.frame(Phyla_fun@sam_data)

## Renombramos las columnas de nuestras abundancias con el phylum de la tabla de taxa
colnames(OTU_matrix) <- Tax_matrix$Phylum 

Fungi_phyla <- cbind(metadata,OTU_matrix)
```

## Normalizacion

```{r}
## shapiro test
for(i in 5:ncol(Fungi_phyla)){
  shapiro <- shapiro.test(Fungi_phyla[,i])
  normal <- ifelse(shapiro[["p.value"]]>0.05, "YES", "NO")
  print(c(i,normal))
}

# solo Ascomycota y olpidiomycota son normales

### Normalizar phylums ###

## Transformacion Log

Norm_Fungi  <- Fungi_phyla

for(i in 6:ncol(Norm_Fungi)){
  Norm_Fungi[,i] <- abs(log10(Norm_Fungi[,i]+1)) # abs es valor absoluto
}

# Volvemos a correr shapiro
for(i in 5:ncol(Norm_Fungi)){
  shapiro <- shapiro.test(Norm_Fungi[,i])
  normal <- ifelse(shapiro[["p.value"]]>0.05, "YES", "NO")
  print(c(i,normal))
}

# Checamos grupos especificos
shapiro.test(Norm_Fungi$p__Mucoromycota) # 6
shapiro.test(Norm_Fungi$p__Basidiomycota) # 7

# concluimos que casi todos nuestros phylums son normales, unos no se pudieron normalizar porque tienen muchos ceros

# Tabla final(solo para datos de isa)

Norm_Fungi <- Norm_Fungi[,-c(5,13,15,18,19,20)] # estoy quitando columna

# 5 y 13 eran normales
# el resto los quitams por baja abundancia

# Unir tablas
Fungi_stats <- cbind(Fungi_phyla[,c(1:5,13)],Norm_Fungi[,5:14]) # seleccionamos las normales de Fungi phyla y lo unimos con las normalizadas de norm fungi

```

## ANOVA de dos vias

```{r}
## Levene test ##
for(i in 5:ncol(Fungi_stats)){
  Lev_ex <- leveneTest(Fungi_stats[,i] ~ Plant_Type * Treatment,
                       data = Fungi_stats)
  levene <- ifelse(Lev_ex[["Pr(>F)"]]>0.05, "YES", "NO")
  print(c(i,levene))
}

# todos pasan levene

## for loop de ANOVA ##

# Paso 1:Tabla vacia
phyla_pvalues <- data.frame(Plant = rep(NA,12), 
                            Treatment = rep(NA,12),
                            Pl_Trt = rep(NA,12)) # el num de variables


for(i in 5:ncol(Fungi_stats)){
  Phyla <- Fungi_stats[,i] # objeto de columna porque Anova no lee la columna
  AR <- Anova(aov(Phyla ~ Plant_Type * Treatment,
                  data = Fungi_stats)) # codigo de ANOVA en un objeto
  j <- i-4 # esto es para que empiece a llenar desde la fila 1 formula general es i-x=1
  phyla_pvalues$Plant[j] <- AR$`Pr(>F)`[1]
  phyla_pvalues$Treatment[j] <- AR$`Pr(>F)`[2]
  phyla_pvalues$Pl_Trt[j] <- AR$`Pr(>F)`[3]
}

# elementos a cambiar cada que adapten este for loop

# 1. Fungi_stats por su tabla de datos
# 2. El numero (5) donde empieza a contar, el numero despues de sus metadatos
# 3. Dentro de la ANOVA grupos o variables independientes o los tratamientos
# 4. j es la formula general es i-x=1. Cambiar el valor de x que en este caso es 4
# 5. Nombres de phyla_pvalues dependeran de sus datos y variables

row.names(phyla_pvalues) <- colnames(Fungi_stats[5:16])

# ueden guardar esta tabla
write.csv(phyla_pvalues, "~/CursoR/CursoRgit/Materiales/Fungi_phyla_pvalues.csv")
```

Baados en estos resultados nosotros podemos elegir cual de estos phylums se puede ver a un nivel taxonomico mas bajo. Asi mismo no tenemos que sacar tukey para todos solo para los significativos:

```{r}
## Mucoromycota, Chytridiomycota y Zoopagomycota

TukeyHSD(aov(p__Zoopagomycota ~ Plant_Type * Treatment,
             data = Fungi_stats))
```

El siguiente paso seria graficar algunas de estas relaciones interesantes

## Ejercicio

Este ejercicio lo haremos juntos. Si bien podria darles el codigo, quise que vieramos juntos esta vez como se podria adpatar un codigo que ya sabemos (en este caso el for loop de ANOVA) a un caso que no sabemos (T test). Si recuerdan los datos de vid no podemos analiarlos por ANOVA porque solo tenemos una variable de cada "grupo". Sin embargo, l hacer la diversidad alfa vimos que si juntamos los datos podemos hacer t tests. En este caso tenemos muchos datos (mas que solo los 7 indices de diversidad alfa) por lo que hacerlo uno por uno seria una tarea ridicula asi que vamo a hacer el analisis usando un for loop:


```{r}
## Preprocesamiento:
Phyla_fun <- tax_glom(vid_bio, taxrank = "Phylum", NArm = FALSE)

## Extraer datos del objeto phyloseq

OTU_matrix <- as.data.frame(Phyla_fun@otu_table)
Tax_matrix <- as.data.frame(Phyla_fun@tax_table)

## Extract metadata
metadata <- as.data.frame(Phyla_fun@sam_data)

## Renombramos las columnas de nuestras abundancias con el phylum de la tabla de taxa
colnames(OTU_matrix) <- Tax_matrix$Phylum 
OTU_matrix <- OTU_matrix[,-c(10,21)]

vid_phyla <- cbind(metadata,OTU_matrix)

## Normalizacion:

for(i in 3:ncol(vid_phyla)){
  shapiro <- shapiro.test(vid_phyla[,i])
  normal <- ifelse(shapiro[["p.value"]]>0.05, "YES", "NO")
  print(c(i,normal))
}

#  10,15,24:28,30:34,37:40
# (vamos a quitar el 10, 15, 27 por baja abundancia) 38:40(estos los quitamos por baja abundancia)

# Seleccionamos tabla
Norm_vid <- vid_phyla[,c(24:26,28,30:34,37)]

# Transformamos con log
for(i in 1:ncol(Norm_vid)){
  Norm_vid[,i] <- abs(log10(Norm_vid[,i]+1)) 
}

for(i in 1:ncol(Norm_vid)){
  shapiro <- shapiro.test(Norm_vid[,i])
  normal <- ifelse(shapiro[["p.value"]]>0.05, "YES", "NO")
  print(c(i,normal))
}

# Juntar para tabla final

vid_stats <- cbind(vid_phyla[,-c(10,15,24:28,30:34,37:40)],Norm_vid) # contando las que eliminamos tendriamos que tener 34 variables

### For loop de t.test ###

# Paso 1. Tabla vacia
phyla_pvalues <- data.frame(Tratamiento = rep(NA, 32),
                            Suelo = rep(NA,32)) # contar el numero de variables

# vamos hacer dos t tests uno para cada variable por eso dos columnas

# Paso 1.5:
objeto_T <- t.test(Proteobacteria ~ Tratamiento, data = vid_stats)
# De aqui vimos como llamar al p value
# objeto_T[["p.value"]]

# ver si se pueden meter valores por columnas:
 t.test(vid_stats[,3] ~ Tratamiento, data = vid_stats)

# Paso 2: for loop
for(i in 3:ncol(vid_stats)){
  T_trat <- t.test(vid_stats[,i]~Tratamiento, data=vid_stats)
  S_trat <- t.test(vid_stats[,i]~Suelo, data=vid_stats)
  j <- i-2 # esto es para que empiece a llenar desde la fila 1 
  phyla_pvalues$Tratamiento[j] <- T_trat[["p.value"]]
  phyla_pvalues$Suelo[j] <- S_trat[["p.value"]]
}
 
row.names(phyla_pvalues) <- colnames(vid_stats[3:34])

# Se guarda esta tabla
write.csv(phyla_pvalues, "~/CursoR/CursoRgit/Materiales/vid_phyla_pvalues.csv")
```

# Conclusion de analisis a nivel phylum

Despues de todos estos graficos y analisis estadisticos ustedes deberian tener unos 5 phylums tops que quieran analizar mas a profundidad. En mi caso me voy a enfocar en Ascomycota y Mucoromycota.

En el caso de vid: 
Cyanobacteria y Proteobacteria basados en graficos y basados en analisis estadistico tenemos Chloroflexi y Actinobacteriota

```{r}
Cyano <- subset_taxa(vid_bio, Phylum == "Cyanobacteria")

Chloro <- subset_taxa(vid_bio, Phylum == "Chloroflexi")
Proteo <- subset_taxa(vid_bio, Phylum == "Proteobacteria")
```

# 2. Visualiar a nivel de orden

Si bien despues de phylum viene clase, por razones de tiemppo(tanto de la clase como de su trabajo) les recomiendo que el siguiente nivel a analizar sea orden:

```{r}
Cyano_ord <- tax_glom(Cyano,taxrank = "Order", NArm = FALSE) # juntar orden

Chloro_ord <- tax_glom(Chloro,taxrank = "Order", NArm = FALSE)
Proteo_ord <- tax_glom(Proteo,taxrank = "Order", NArm = FALSE)
```

En este caso no lo hice para Mucoromycota tambien porque si se fijan solo tiene un orden. Aglomerar todos nuestros datos por orden no nos serviria de nada por lo que les vuelvo a reiterar que aqui depende de cada caso para que sigan su flujo de analisis. A partir de aqui yo tendria 2 opciones: 

1. Si se fijan a nivel de familias si habian varias clasificaciones por lo que podria analizar a ese nivel en vez de a nivel de orden o

2. Pasarme directamente al siguiente nivel de analisis: el genero

Nuevamente al ver los datos (sin graficar) me doy cuenta que las familias y los generos siguen los mismos patrones por lo que en vez de analizar 2 veces dejaremos Mucoromycota para el siguiente nivel.

### Heatmap o barplot?

Primero vamos a ver el barplot default sin arreglar nada en los datos:

```{r}
plot_composition(Chloro_ord, plot.type = "barplot")+ 
  theme_bw()+
  ggtitle("Abundancias de Chloroflexi a nivel Orden")+
  theme(plot.title = element_text(hjust = 0.5, size = 15, face = "bold"))+
  xlab("Muestras")+
  ylab("Abundancia")
# Nos fuimos por el grupo mas abundante

# IQHNW1012T
# IJISH7931T
# VPLBK6313M

plot_composition(Proteo_ord, plot.type = "barplot")+ 
  theme_bw()+
  ggtitle("Abundancias de Proteobacteria a nivel Orden")+
  theme(plot.title = element_text(hjust = 0.5, size = 15, face = "bold"))+
  xlab("Muestras")+
  ylab("Abundancia")
```

Anteriormente les habia dicho que la funcion plot_composition de microbiome tiene tambien una opcion de heatmap. El problema es que es mucho mas tardado de correr y los graficos no quedan tan bien. L otra opcion es la de phyloseq que aunque corre mejor en mi opinion no es tan sensible a cambios como pheatmap.

Sin embargo, en este caso no vamos a pasar por todo el proceso de extraer tablas y hacer un "buen" heatmap por una simple razon: A pesar de que si nos estamos tomando el tiempo de analizar orden yo lo veo mas como un nivel de transicion. Es decir nos ayuda a dirigir nuestro analisis a el o a los generos que podrian llegar a ser interesantes asi que les voy a ensenar dos opciones de heatmap rapidas pero feas:

```{r}
# Heatmap con microbiome
plot_composition(Chloro_ord, plot.type = "heatmap")+
  theme_bw()+
  coord_flip()
# Si tenemos muhas muestras es mejor usar barplot porque pude morir la compu

# Usando la funcion de phyloseq
plot_heatmap(Chloro_ord, 
             low = "#D8AFFF", high = "#320743",
             na.value = "snow2")

plot_heatmap(Proteo_ord, 
             low = "lightskyblue", high = "midnightblue",
             na.value = "snow2")


```

Los graficos a nivel de orden son exploraorios y por lo mismo decidi solo mostrarles como se verian con funciones basicas.

Aunque en mi caso es trampa(porque yo ya se como se comportan mis datos), de ambos graficos yo podria concluir que el rden de mayor interes es Hypocreales ya que tiene mucha mas abundancia que el resto. Sin embargo i no ven un patron exacto o hay mucha variacion en sus datos si podrian meter sus datos en uno de los forloop que ya hemos creado.

Conclusiones de vid a nivel de orden:

1. Chloroflexi: por diferencias

* Thermomicrobiales: 
* Chloroflexales: 
* SBR1031: 

Por analisis estadistico

* Thermomicrobiales: esta al margen de diferencias estadisticas
* SBR1031: si presento diferencias estadisticas

2. Proteobacteria: por abundancia

* Rhizobiales

Por analisis estadistico:

* Caulobacterales, Sphingomonadales: presentaron diferencias estadisticas

3. Actinobacteria: casi uniforme

* Micromonosporales (por abundancia)
* Propionibacteriales 
* Micrococcales 
* Streptomycetales

Por analisis estadistico:

* Corynecbacteriales: si presento dif estadisticas

4. Cyanobacteria: por abundancia

* Chloroplast

Por analisis estadistico:
No hubieron diferencias estadisticas

## Ejercicio: Analisis estadistico

Usando sus datos de vid pasen los valores de orden para determinar el camino de analisis adaptando los forloops de t. test que creamos para phylum.

```{r}

OTU_matrix <- as.data.frame(Chloro_ord@otu_table)
Tax_matrix <- as.data.frame(Chloro_ord@tax_table)

metadata <- as.data.frame(Chloro_ord@sam_data)

## Renombramos las columnas de nuestras abundancias con el phylum de la tabla de taxa
colnames(OTU_matrix) <- Tax_matrix$Order 

vid_phyla <- cbind(metadata,OTU_matrix)

## Revisar normalizacion:

for(i in 3:ncol(vid_phyla)){
  shapiro <- shapiro.test(vid_phyla[,i])
  normal <- ifelse(shapiro[["p.value"]]>0.05, "YES", "NO")
  print(c(i,normal))
}

order_pvalues <- data.frame(Tratamiento = rep(NA,23),
                            Suelo = rep(NA,23))

# for loop
for(i in 3:ncol(vid_phyla)){
  T_trat <- t.test(vid_phyla[,i]~Tratamiento, data=vid_phyla)
  S_trat <- t.test(vid_phyla[,i]~Suelo, data=vid_phyla)
  j <- i-2 # esto es para que empiece a llenar desde la fila 1 
  order_pvalues$Tratamiento[j] <- T_trat[["p.value"]]
  order_pvalues$Suelo[j] <- S_trat[["p.value"]]
}

row.names(order_pvalues) <- c( "NA1", "S085","SBR1031","Ardenticatenales","Anaerolineales", "NA2","SAR202 clade","NA4", "RBG-13-54-9","NA5","NA6","C0119","NA7","Chloroflexales","NA8", "NA9","Caldilineales","NA10","Thermomicrobiales","Kallotenuales", "NA11",        "Ktedonobacterales","NA12")

write.csv(order_pvalues, "~/CursoR/CursoRgit/Materiales/vid_order_pvalues.csv")



OTU_matrix <- as.data.frame(Proteo_ord@otu_table)
Tax_matrix <- as.data.frame(Proteo_ord@tax_table)

metadata <- as.data.frame(Proteo_ord@sam_data)

## Renombramos las columnas de nuestras abundancias con el phylum de la tabla de taxa
colnames(OTU_matrix) <- Tax_matrix$Order 

vid_phyla <- cbind(metadata,OTU_matrix)

## Revisar normalizacion:

for(i in 3:ncol(vid_phyla)){
  shapiro <- shapiro.test(vid_phyla[,i])
  normal <- ifelse(shapiro[["p.value"]]>0.05, "YES", "NO")
  print(c(i,normal))
}

order_pvalues <- data.frame(Tratamiento = rep(NA,23),
                            Suelo = rep(NA,23))

# for loop
for(i in 3:ncol(vid_phyla)){
  T_trat <- t.test(vid_phyla[,i]~Tratamiento, data=vid_phyla)
  S_trat <- t.test(vid_phyla[,i]~Suelo, data=vid_phyla)
  j <- i-2 # esto es para que empiece a llenar desde la fila 1 
  order_pvalues$Tratamiento[j] <- T_trat[["p.value"]]
  order_pvalues$Suelo[j] <- S_trat[["p.value"]]
}

row.names(order_pvalues) <- c( "NA1", "S085","SBR1031","Ardenticatenales","Anaerolineales", "NA2","SAR202 clade","NA4", "RBG-13-54-9","NA5","NA6","C0119","NA7","Chloroflexales","NA8", "NA9","Caldilineales","NA10","Thermomicrobiales","Kallotenuales", "NA11",        "Ktedonobacterales","NA12")

write.csv(order_pvalues, "~/CursoR/CursoRgit/Materiales/vid_order_pvalues.csv")
```

Dentro del orden u ordenes que eleigieron se puede seguir el proceso de filtrado y aglomeracion anterior para ir eligiendo un genero (en el caso de bacterias) o algunas especies en el caso de hongos (para analisis).

Si bien a mi como ecologa de suelo se me hace interesante ver si realmente hay esta clase de cambios a niveles mas bajos y luego como estos se traducen a cambios funcionales a estas alturas su reporte va a ser del largo de mi tesis. Por ello en esta seccion solo nos vamos a enfocar en ciertos generos y especies que son importantes reportar ya que son conocidos por su importancia tanto buena como mala.

A menos que les pidan un analisis super minucioso, yo les recomiendo que hagan el analisis de phylum para darse una idea global de los datos y de ahi se enfoquen en las siguientes especies como marcadores de salud de suelo:

__Bacterias:__

* Pseudomonas 67

* Rhizobium 0

* Rubrobacter (opcional) 7

* Rhizobacter 0

* Clavibacter 0

* Streptomyces 132

* Bacillus 41

__Hongos:__

* Trichoderma           

* Glomus

* Fusarium

* Apophysomyces 

En my caso solo veremos los graficos de Fusarium y Apophysomyces

```{r}
Rubro <- subset_taxa(vid_bio, Genus == "Rubrobacter")
```

Si estan a nivel de genero(es decir con bacteria) ustedes juntarian todos los conteos de dicho genero graficarian y verian si vale la pena o no hacer el analisis estadistico

En el caso de hongos(es decir el ejemplo a continuacion) primero tengo que juntar por especie ya que la mayoria fueron identificadas. De ahi lo que yo hice fue repetir los for loops de analisis estadistico para ver si alguna o todas las especies mostraban diferencias estadisticas. Este proceso es un poco mas tardado pero ya tenemos todas las herramientas para hacerlo.

## Pre-procesamiento de datos

```{r}
Fusarium <- tax_glom(Fusa, taxrank = "Species", NArm= FALSE)

# Extraer matriz
OTU_matrix <- as.data.frame(Fusarium@otu_table)
Tax_matrix <- as.data.frame(Fusarium@tax_table)

# Renombrar las columnas de otumatrix con el phylum de taxmatrix
colnames(OTU_matrix) <- Tax_matrix$Species 

metadata <- as.data.frame(Fusarium@sam_data)

# Unir phyla con metadata
Fusarium <- cbind(metadata,OTU_matrix)

# Suma total de Fusarium
Sum_fus <- rowSums(Fusarium[,5:14])
Sum_fus <- cbind(metadata,Sum_fus)

# Shapiro test
shapiro.test(Sum_fus$Sum_fus) # normal

# Levene test
leveneTest(Sum_fus ~ Plant_Type * Treatment,
           data = Sum_fus) # si lo paso el test

# ANOVA
Anova(aov(Sum_fus ~ Plant_Type * Treatment,
          data = Sum_fus)) # no significativo

```

# Especies especificas

Despues de hacer el analisis anterior para cada una de las especies descubri 2
coas _F. tricinctum_ mostraba diferencias estadisticas entre tipos de planta (1ue no se los voy a mostrar) mientras que _F. verticillioides_ mostraba diferencia por tratamiento:

```{r}
ggplot(Fusarium, aes(x = Treatment, y =s_verticillioides, col = Plant_Type))+
  geom_boxplot(fill = "snow2", lwd = 0.7)+
  theme_bw(base_size=15)+
  scale_color_manual(name = "Plant Type",
                     values =c("steelblue", "lightpink","burlywood3"))+
  xlab("Treatment") +
  ylab("Abundance")+
  geom_signif(y_position = 880, xmin=0.6, xmax=3.4, annotation="A", tip_length = 0.01, col=1)+
  geom_signif(y_position = 1100, xmin=3.6, xmax=4.4, annotation="B", tip_length = 0.01, col=1)
```

### Ejercicio 

Graficar su genero en especifico

```{r}
Bacillus <- subset_taxa(vid_bio, Genus == "Bacillus")

# Extraer matriz
OTU_matrix <- as.data.frame(Bacillus@otu_table)
metadata <- as.data.frame(Bacillus@sam_data)

Bacillus <- cbind(metadata,OTU_matrix)

# Opcion 1 para grafico de barras (no lo vamos hacer)
Sum_b <- rowSums(Bacillus[,3:9])
Sum_bac <- cbind(metadata,Sum_b)

# para boxplot
Bacillus_m <- reshape2::melt(Bacillus)

ggplot(Bacillus_m, aes(x = Tratamiento, y =value))+
  geom_boxplot(lwd = 0.7)+
  theme_bw(base_size=15)+
  ggtitle("Abundancia en el suelo de Bacillus")+
  theme(plot.title = element_text(hjust = 0.5, size = 15, face = "bold"))+
  scale_fill_manual(values =c("#00CD66", "steelblue4"))+
  xlab("Tratamiento") +
  ylab("Abundancia")


Streptomyces <- subset_taxa(vid_bio, Genus == "Streptomyces")

# Extraer matriz
OTU_matrix <- as.data.frame(Streptomyces@otu_table)
metadata <- as.data.frame(Streptomyces@sam_data)

Streptomyces <- cbind(metadata,OTU_matrix)

# Opcion 1 para grafico de barras (no lo vamos hacer)
Sum_s <- rowSums(Streptomyces[,3:9])
Sum_st <- cbind(metadata,Sum_s)

# para boxplot
St_m <- reshape2::melt(Streptomyces)

ggplot(St_m, aes(x = Tratamiento, y =value, col = Suelo))+
  geom_boxplot(lwd = 0.7)+
  theme_bw(base_size=15)+
  xlab("Tratamiento") +
  ylab("Abundancia")
  
  
```

 # Conclusiones de analisis taxonomico
 
 Como pueden ver el analisis taxonomico es muy exploratorio, se pueden tardar 1 semana o una hora dependiendo el nivel de profundidad que le quieran dar. De cajon les recomiendo que chequen phylum y nivel de genero analizar los grupos que les mostre para todos sus reportes. Muchas veces el analisis a niveles taxonomicos altos los puede llevar a analizar otro generos que tambien son de relevancia agricola pero como pueden ver ir de gran a pequeña escala es un proceso un poco tardado aunque no es dificil. La ventaja de que lo hagan entre todos es que cada quien
