==> Notebook/20_Analisis_hongos.Rmd <==
---
title: "Analisis funcional para hongos"
output: html_notebook
---

# Introduccion

Desafortunadamente , Metagenassist olo funciona para la asignacion funcional de fenotipos en bacterias por l que para hongos usaremos la base de datos FungalTraits. FungalTraits es una base de datos de rasgos de hongos acoplada para ser usada en RStudio. La base fue creada combinando informacion de bsases de datos previas conocidas como FUNGuild y Fun las cuales usaninformacion de UNITE con respecto a caracteristicas fenotipicas de cientos de secuencias de hongos a nivel de genero y especie.

```{r}
# Instalacion
devtools::install_github("ropenscilabs/datastorr")
devtools::install_github("traitecoevo/fungaltraits")

# Librerias
library(htmltools); packageVersion("htmltools") # html tools solo se puede actualizar si todos los Rmarkdowns estan abiertos

library(fungaltraits) # es la base de datos
library(phyloseq)
library(tidyverse)
library(dplyr)
library(microeco)
library(pheatmap)
library(scales)
library(ggsignif)

fungal_traits()
```

# Pre-procesamiento de datos

```{r}
load("ITS_filt.RData")

ITS_genus <- subset_taxa(ITS_filtered, !is.na(Genus))

### Extraer tax table

Fun_genus <- as.data.frame(ITS_genus@tax_table)
write.csv(Fun_genus, "~/CursoR/CursoRgit/Materiales/Fungal_genus.csv")
# en excel vamos a quitar lo de g__ y s__ para que los datos puedan ser unidos
# se quita con reemplazar y filtrar

# Subir tabla corregida

Fun_genus <- read.csv("~/CursoR/CursoRgit/Materiales/Fungal_genus.csv")

#row.names(Fun_genus) <- Fun_genus[,1]
#Fun_genus <- Fun_genus[,-1]

# Crear objetos con las tablas de a base de datos

Fungal_traits <- fungal_traits() # original
FT <- fungi_func_FungalTraits # de datos suplementarios usando microeco
FUNGuild <- fungi_func_FUNGuild # comparar con base de datos antiguas usando microeco
```

# Asignacion de grupos funcionales

En este caso el proceso de asignacion va a ser "manual" usando el codigo de R merge(). Para ello se usa una varible comun entre dos tablas y la mas especifica es genero:

```{r}
# Renombrar columna Genus
colnames(FT)[6] = "Genus"

## Merge
Fun_fun <- merge(x = Fun_genus, y = FT, by = "Genus")

### Limpiar la tabla
Fun_fun <- Fun_fun[,-(3:7)] 
Fun_fun <- Fun_fun[,-(4:9)] 
Fun_fun <- Fun_fun[,-6] 


# Remover columnas vacias
empty_columns <- sapply(Fun_fun, function(x) all(is.na(x) | x == ""))

Fun_fun <- Fun_fun[,-(18:19)]

# Guardar la tabla
write.csv(Fun_fun, "~/CursoR/CursoRgit/Materiales/Fungi_funct.csv")
```

En esta tabla unimos la identificacion taxonomica con la categoria fenotipica de "guild" que se traduce como nicho, es decir el rol que potencialmente podrian cumplir en el ecosistema. El siguiente paso es unir estos datos con la tabla de abundancia para ver cuantos hongos cumplen cada rol:

```{r}
Fun_counts <- as.data.frame(ITS_genus@otu_table) # extraer de phyloseq

Fun_counts <- Fun_counts[,Fun_fun[,2]] # filtrar estilos de vida identificados

## Unir el estilo de vida y secundario a la tabla de cuentas

Fun_counts[nrow(Fun_counts)+1,] <- Fun_fun[,4]
Fun_counts[nrow(Fun_counts)+1,] <- Fun_fun[,5]

# Guardar esta tabla
write.csv(Fun_counts, "~/CursoR/CursoRgit/Materiales/Fungi_lifestyles.csv")
```

Esta tabla final la modifique en excel para unir las cuentas de cada grupo. Esto tambien se puede hacer en R usando funciones de dplyr:

```{r}
Fun_counts %>%
  mutate(soilsapro_sum = colSums(1:20))
```

### Preprocesamiento de datos

```{r}
Simp_fungi <- read.csv("~/CursoR/CursoRgit/Materiales/Simplified_Fungi.csv")

row.names(Simp_fungi) <- Simp_fungi[,1]
Simp_fungi <- Simp_fungi[,-1]

# Siempre cambiar a matriz
Simp_fungi <- data.matrix(Simp_fungi)

Simp_fungi <- Simp_fungi[order(Simp_fungi[,1],
                               decreasing = TRUE),]

sorder <- c("A1","B1","C1","A2","B2","C2", 
            "A3","B3","C3", "A4","B4","C4", 
            "A5","B5","C5", 
            "A6","B6","C6", "A7","B7","C7", 
            "D1","D2","D3","D4")

Simp_fungi <- Simp_fungi[, sorder]

## Quantile breaks
quantile_breaks <- function(xs, n = 10) {
  breaks <- quantile(xs, probs = seq(0, 1, length.out = n))
  breaks[!duplicated(breaks)]
} 

# breaks object 10 breaks
mat_breaks <- quantile_breaks(Simp_fungi, # es la matriz
                              n = 10)

# Heatmap
pheatmap(Simp_fungi,
         cluster_rows = FALSE, cluster_cols = FALSE,
         scale = "none", 
         breaks = mat_breaks,
         color = colorRampPalette(c("snow2", "#B6AAE1", "mediumpurple3", "#0E021A"))(9),
         fontsize = 12,
         gaps_col = c(3,6,9,12,15,18,21),
                  labels_col = c (labels=c(rep("36LP",3),rep("WTLP",3),  
                                  rep("36Pi",3), rep("WTPi",3),  
                                  rep("36Phi",3), 
                                  rep("36PM",3), rep("WTPM",3),
                                  "BSLP","BSPi","BSPhi","BSPM")))

```
## Grupos especificos

Si bien recuerdan FungalTraits  nos da un estilo de vida primario (que es el que usamos hace rato para el heatmap) pero tambien definio algunas de las especies con un estilo de vida secundario. Uno de estos grupos dentro de esta categoria es especies asociadas a raices (root-associated) usando esa clasificacion, graficaremos los hongos que ya han sido especificamente clasificados como habitantes de la rizosfera:

### Preprocesamiento: root-associated

```{r}
Func_Fungi <- read.csv("~/CursoR/CursoRgit/Materiales/Functional_Fungi.csv")

metadata <- read.csv("~/CursoR/CursoRgit/Materiales/metadata.csv")

### Preprocesamiento

Func_Fungi <- as.data.frame(t(Func_Fungi))
colnames(Func_Fungi) <- Func_Fungi[1,]
Func_Fungi <- Func_Fungi[-1,]

## Cambiar a datos numericos
for(i in 1:ncol(Func_Fungi)){
  Func_Fungi[,i] <- as.numeric(Func_Fungi[,i])
}

# Seleccionar root-associated
Root_sums <- rowSums(Func_Fungi[,c(11,12,22,36,40,48,50,53,55,63,66,73,77,81,86,88,94)]) # sumo

Root_sums <- cbind(metadata,Root_sums)

# Remover controles
Root_cat <- Root_sums[c(1:21),]

```

### Boxplot

```{r}
# Agregamos una fila vacia para que el grafico quede simetrico
Root_cat[nrow(Root_cat)+ 1,] <- list("E1", "WTPhi","Wild Type", "Phi", NA,0)

ggplot(Root_cat, aes(x = Treatment, y = Root_sums, fill = Plant_Type))+
  geom_boxplot()+
  theme_bw(base_size = 15)+
  scale_fill_manual(name="Plant Type",
                    values = c("steelblue", "lightpink"))+
  xlab("Treatment") + ylab("Abundance")+
  geom_signif(y_position = 7500, xmin=2.6, xmax=3.4,
                annotation = "p<0.01", tip_length = 0.01,
                col=1)+
  geom_signif(y_position = 4100, xmin=0.6, xmax=1.4,
                annotation = "NS", tip_length = 0.01,
                col=1)+
  geom_signif(y_position = 5000, xmin=3.6, xmax=4.4,
                annotation = "NS", tip_length = 0.01,
                col=1)+
  scale_y_continuous(breaks=pretty_breaks(n=6))
  
```
